Commented some of SwerveMath. Added experimental fix in comments
Our driver was saying that he didn't like the wheels always snapping to zero, this is fixable and this is the start of working on it. 
The main problem with this issue was previously solved by Dr. Maples in Deadzone_Correction where the x2 value was saved and always set, even at 0. So the angle it was at before it hit 0 would still be the angle after he let go.
What I have done is solved the math side of this problem where we need to know the difference it needs between the current angle and the wanted angle.

This code is commented out and will require review (by our mentor) before actually deploying it to the bot. It also wouldn't work yet, because I haven't yet found a way to read the angles from the Angle motors.